# Avoid fmt.Sprintf for simple subjects