@nuss-justin 2016-10-05
The error is unrelated and occurs even without this change :-/