@kozlovic 2016-10-05
Thanks. The failure is due to the fact that the server has already the support for durable queue subscribers. There is a PR in this repo to add doc and change the failing test. So please ignore this failure.