@kozlovic 2016-10-05
Otherwise, this looks good to me. I will merge when the other PR is merged so that build on master does not fail. Thanks again!