@kozlovic 2016-10-06
@nuss-justin As a sanity check, could you rebase from master and update the PR to see if the build succeeds now that I have merged the other PR? Thank you!